chore: move namespaced function to util
