+++ lib/client.js
@@ -1,11 +1,12 @@
 'use strict';
 
-const crypto   = require('crypto'),
-      get      = require('./http-then').get,
-      ExecAuth = require('./auth/exec'),
-      GCPAuth  = require('./auth/gcp'),
-      OIDCAuth = require('./auth/oidc'),
-      URI      = require('urijs');
+const crypto     = require('crypto'),
+      get        = require('./http-then').get,
+      ExecAuth   = require('./auth/exec'),
+      GCPAuth    = require('./auth/gcp'),
+      OIDCAuth   = require('./auth/oidc'),
+      namespaced = require('./util').namespaced,
+      URI        = require('urijs');
 
 const { error } = require('./error');
 
@@ -208,18 +209,8 @@ class Client {
     return this._paths.some(path => path === '/apis/apps.openshift.io');
   }
 
-  // Adds a namespace to the API request if needed
-  namespacedPath(strings, namespace) {
-    const path = strings[0];  
-    if (namespace === undefined) {
-      return path;
-    }
-    const i = path.lastIndexOf('/'); 
-    return path.slice(0, i) + '/namespaces/' + namespace + path.slice(i, path.length);
-  }
-
   events(namespace) {
-    return new WatchableRequest({ path: this.namespacedPath`/api/v1/events?limit=500${namespace}`, headers: { 'Accept' : 'application/json;as=Table;v=v1beta1;g=meta.k8s.io, application/json' }}, { client: this });
+    return new WatchableRequest({ path: namespaced`/api/v1/events?limit=500${namespace}`, headers: { 'Accept' : 'application/json;as=Table;v=v1beta1;g=meta.k8s.io, application/json' }}, { client: this });
   }
 
   event(namespace, name) {
