+++ lib/util.js
@@ -104,3 +104,13 @@ module.exports.isLocalStorageAvailable = function () {
       storage.length !== 0;
   }
 }
+
+// Adds a namespace to the API request if needed
+module.exports.namespaced = function (strings, namespace) {
+  const path = strings[0];
+  if (namespace === undefined) {
+    return path;
+  }
+  const i = path.lastIndexOf('/');
+  return path.slice(0, i) + '/namespaces/' + namespace + path.slice(i, path.length);
+}
